Fix calculation of the size of our signature data
The 'size' field of the certificate table header includes the size of
the header itself.  When parsing a signed file, we should therefore
subtract the size of this header from the field representing the size of
the pkcs7 data packet; otherwise when we detach (and subsequently
reattach) a signature, we wind up with 8 extra bytes of zeroes at the
end each time.  Fixing this ensures that detaching and signature and
then reattaching it to the file gives us back the original file.